Organized migration files. Merge some audio utilities.
